requirements.txt
flask==3.0.3
python-dotenv
flask-WTF==1.2.1
flask-Login==0.6.3
pymongo==4.8.0
email-validator==2.2.0
qrcode[pil]==7.4.2
Pillow==10.4.0
stripe>=10.0.0
paho-mqtt==1.6.1
itsdangerous>=2.1  # pour signer les QRcode
